Reorganize files for better sharing
Matching changes for sharing after https://github.com/dotnet/runtime/pull/47323.
